refactor(admin): consolidate feature flags into admin console as a tab
- Replace separate /admin/feature-flags page with a second tab inside the admin console
- Remove Feature Flags entry from sidebar (all admin writes live in the console)
- Update page subtitle to reflect multi-section scope
- Add PrimeVue Tabs (TabList / Tab / TabPanels / TabPanel) with Users and Feature Flags panels
- Users panel: identical layout to previous admin page (two-column grid preserved)
- Feature Flags panel: flag list with ToggleSwitch, status tag, empty state, refresh button
- Update admin.spec.ts: add slot-rendering Tab stubs, add tests for flags panel rendering (270 total)
